Provide a custom derivation as "configuration"
This allows to have a list the current installed set of derivations.
Instead of using nix-env like apt.

Use `nix-env -i -f configuration.nix` to install this derivation.
It will then install defined "packages".

This eases to maintain current system and port to NixOS or something
else in the future.