Add first basic flake to be used as flake.nix in projects
The flake is very basic and encapsulates frontend compilation via node
for customer projects.
The flake is intended to be used by other flake.nix within the projects,
which will use the defined packages to build apps outputs.
They should use the packages like `writeShellApplication` itself,
call it with necessary argument set in order to receive a
out of the box working solution.

A template is also provided which can be used like:

    nix flake init -t <path to this flake>

`defaultTemplate` is defined in output and will be used.