Remove runtime inputs for system update script
In order to use from actual path, prevent issues with broken versions.
The script was linked against an 2021 version of home-manager which was
incompatible with current unstable nixpkgs.
That prevented updates.

There is now no concrete version, but it will use the once from current
PATH.
This allows to use the latest "current" version as defined by system.