Configure Neovim completion
Still use menu like default.
Do not show preview as in default.
Do not insert as in default.
Do not select as in default.

This allows me to see suggestions but continue typing to narrow down.
Also allows me to manually select a result.

That way completion is out of my way but there to support me.