Migrate nix and nix registry configuration to home manager module
There is an existing module.
Existing manually written and linked files are replaced with usage of
the module.

This has the advantage that I don't need to know where to place the
files and which format to use.
Furthermore, the module will check whether I've typos in options and
whether values match expected types.