Update todos in readme
system tools are not possible and I stick to ubuntu as a host os.
In general I shouldn't care to much about the host OS.
I use nix for most things and will improve the situation.

Most customers also have ubuntu or debian systems as servers.
It is a good thing to have the same folder and software setup.