Fix broken vim php syntax
The loading order (runtimepath) was changed with
https://github.com/NixOS/nixpkgs/commit/cda1f8ae46869c429971323717d622d5b17d9854.
The php syntax loaded as package is now loaded after the neovim native
php syntax file and does not have any effect.

I now remove the package and add the file as XDG config so it is always
loaded first.
This file is not maintained anymore since some years and I still prefer
it over the maintained versions.